Handle changes in signing key size; remove signature type option.
Changing the size of the signing key results in a signature of a
different length that needs to be dealt with in the top level sequence.

Allowing the user to specify the signature algorithm is pointless since
if what they provide doesn't match the existing CRL, we're doomed.
Providing a parameter that has only one correct setting is useless.  In
the future it might be worth investigating what it would take to allow
changing the signatureAlgorithm.  Based on the existing preprocessing we
have to do, I'm guessing it wouldn't be that much extra work.