Add special case code to handle empty CRLs.
Empty CRLs are missing the revokedCertificates sequence which creates a
difficult situation for the streaming code since it has to consume part
of the stream in order to determine if the list is empty or not.  If the
list isn't empty, then we've consumed part of the ASN1 that we actually
needed.  This state is especially problematic for the
X509CRLStreamWriter because it relies on X509CRLEntryStream to leave the
stream in a known state.